# backend/static/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "DiabeCare API",
    "description": "DiabeCare \u2014 Diabetic Patient Medication Follow-Up System API",
    "version": "2.0.0"
  },
  "servers": [
    {
      "url": "https://your-production-domain.com/api",
      "description": "Production"
    },
    {
      "url": "http://localhost:5000/api",
      "description": "Local dev"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "BearerAuth": []
    }
  ],
  "paths": {
...
    "/appointments/": {
      "get": {
        "tags": [
          "Appointments"
        ],
        "summary": "List appointments \u2014 role-scoped (admins all, doctors their own, patients their own)",
        "responses": {
          "200": {
            "description": "Array of appointment objects"
          }
        }
      },
      "post": {
        "tags": [
          "Appointments"
        ],
        "summary": "Schedule an appointment (admin/doctor)",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "patient_id",
                  "appointment_date"
                ],
                "properties": {
                  "patient_id": {
                    "type": "integer"
                  },
                  "doctor_id": {
                    "type": "integer"
                  },
                  "appointment_date": {
                    "type": "string",
                    "format": "date-time",
                    "example": "2025-07-20T09:00:00"
                  },
                  "type": {
                    "type": "string",
                    "example": "Follow-up"
                  },
                  "notes": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Appointment scheduled"
          },
          "403": {
            "description": "Admin/doctor access required"
          }
        }
      }
    },
    "/appointments/request": {
      "post": {
        "tags": [
          "Appointments"
        ],
        "summary": "Patient requests an appointment (patient only)",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "requested_date",
                  "type"
                ],
                "properties": {
                  "requested_date": {
                    "type": "string",
                    "format": "date-time",
                    "example": "2025-08-05T10:00:00"
                  },
                  "type": {
                    "type": "string",
                    "example": "Routine Check"
                  },
                  "notes": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Appointment request submitted with status=requested"
          },
          "403": {
            "description": "Patient access only"
          }
        }
      }
    },
    "/appointments/{appt_id}": {
      "put": {
        "tags": [
          "Appointments"
        ],
        "summary": "Update appointment (admin/doctor)",
        "parameters": [
          {
            "name": "appt_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "appointment_date": {
                    "type": "string",
                    "format": "date-time"
                  },
                  "type": {
                    "type": "string"
                  },
                  "status": {
                    "type": "string",
                    "enum": [
                      "scheduled",
                      "completed",
                      "cancelled",
                      "requested",
                      "missed"
                    ]
                  },
                  "notes": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Appointment updated"
          },
          "404": {
            "description": "Appointment not found"
          }
        }
      },
      "delete": {
        "tags": [
          "Appointments"
        ],
        "summary": "Delete appointment (admin/doctor)",
        "parameters": [
          {
            "name": "appt_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Appointment deleted"
          },
          "403": {
            "description": "Admin/doctor access required"
          },
          "404": {
            "description": "Appointment not found"
          }
        }
      }
    },
...
  },
...
}